refactor: add index file
