+++ YuJin/react-ts/src/app/constants/diaryOutputs.ts
