+++ YuJin/react-ts/src/app/emotions/page.tsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom'
 import { EMOTION_LIST } from '../constants/emotions'
 import { emotionColorVariants } from '../../styles/emotionCard'
-import { EMOTION_LIST_DISCRIPTION, EMOTION_LIST_TITLE } from '../constants/diaryOutput'
+import { EMOTION_LIST_DISCRIPTION, EMOTION_LIST_TITLE } from '../constants/diaryOutputs'
 import { ROUTE_TYPE } from '../constants/routes'
 export default function EmotionLinkPage() {
     return (
