--- YuJin/react-ts/src/interface/diaryDetail.interface.ts
@@ -1,3 +1,0 @@
-type DiaryDetailPageParams = {
-    id: string
-}
